refactor: UserLeague 도메인 아키텍처 리팩토링
- infrastructure 패키지 제거, repository 패키지 추가

- League 랭킹 조회, 내 랭크 조회 관련 controller, service 클래스 단일 컨트롤러, 서비스 클래스로 수정

- League 랭킹 조회, 내 랭크 조회 개별 repository 를 UserLeagueRepository에서 상속
