+++ src/main/java/gravit/code/userLeague/controller/UserLeagueController.java
@@ -1,11 +1,11 @@
 package gravit.code.userLeague.controller;
 
-
 import gravit.code.auth.domain.LoginUser;
 import gravit.code.global.dto.response.SliceResponse;
-import gravit.code.userLeague.controller.docs.LeagueRankingQueryControllerDocs;
+import gravit.code.userLeague.controller.docs.UserLeagueControllerDocs;
 import gravit.code.userLeague.dto.response.LeagueRankRow;
-import gravit.code.userLeague.service.LeagueRankingQueryService;
+import gravit.code.userLeague.dto.response.MyLeagueRankWithProfileResponse;
+import gravit.code.userLeague.service.UserLeagueQueryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -19,15 +19,24 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/api/v1/ranking")
 @RequiredArgsConstructor
-public class LeagueRankingQueryController implements LeagueRankingQueryControllerDocs {
-    private final LeagueRankingQueryService leagueRankingQueryService;
+public class UserLeagueController implements UserLeagueControllerDocs {
+
+    private final UserLeagueQueryService userLeagueQueryService;
+
+    @GetMapping("/me")
+    public ResponseEntity<MyLeagueRankWithProfileResponse> getMyLeagueWithProfile(@AuthenticationPrincipal LoginUser loginUser) {
+        Long userId = loginUser.getId();
+        MyLeagueRankWithProfileResponse myLeagueRankWithProfile = userLeagueQueryService.getMyLeagueRankWithProfile(userId);
+
+        return ResponseEntity.ok(myLeagueRankWithProfile);
+    }
 
     @GetMapping("/leagues/{leagueId}/page/{pageNum}")
     public ResponseEntity<SliceResponse<LeagueRankRow>> getLeagueRanking(
             @PathVariable("leagueId") Long leagueId,
             @PathVariable("pageNum") int pageNum
     ){
-        SliceResponse<LeagueRankRow> sliceResponse = leagueRankingQueryService.findLeagueRanking(leagueId, pageNum);
+        SliceResponse<LeagueRankRow> sliceResponse = userLeagueQueryService.findLeagueRanking(leagueId, pageNum);
         return ResponseEntity.ok(sliceResponse);
     }
 
@@ -37,7 +46,7 @@ public class LeagueRankingQueryController implements LeagueRankingQueryControlle
             @AuthenticationPrincipal LoginUser loginUser
     ){
         Long userId = loginUser.getId();
-        SliceResponse<LeagueRankRow> sliceResponse = leagueRankingQueryService.findLeagueRankingByUser(userId, pageNum);
+        SliceResponse<LeagueRankRow> sliceResponse = userLeagueQueryService.findLeagueRankingByUser(userId, pageNum);
         return ResponseEntity.ok(sliceResponse);
     }
 }
