--- src/main/java/gravit/code/userLeague/domain/UserLeagueRepository.java
@@ -1,23 +1,0 @@
-package gravit.code.userLeague.domain;
-
-import gravit.code.league.domain.League;
-import gravit.code.season.domain.Season;
-
-import java.util.Optional;
-
-public interface UserLeagueRepository {
-
-    Optional<String> findUserLeagueNameByUserId(long userId);
-
-    boolean existsByUserId(long userId);
-
-    void save(UserLeague userLeague);
-
-    Optional<UserLeague> findByUserId(long userId);
-
-    int resetAllForNextSeason(
-            Season currentSeason,
-            Season nextSeason,
-            League startLeague
-    );
-}
