+++ src/main/java/gravit/code/userLeague/repository/custom/JdbMyLeagueProfileQueryRepository.java
@@ -1,14 +1,13 @@
-package gravit.code.userLeague.infrastructure;
+package gravit.code.userLeague.repository.custom;
 
 import gravit.code.userLeague.dto.response.MyLeagueRankWithProfileResponse;
-import gravit.code.userLeague.service.MyLeagueProfileQueryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 import java.util.Map;
 
-import static gravit.code.userLeague.infrastructure.sql.MyLeagueRankWithProfileQuerySql.FIND_MY_RANK_WITH_PROFILE_SQL;
+import static gravit.code.userLeague.repository.sql.MyLeagueRankWithProfileQuerySql.FIND_MY_RANK_WITH_PROFILE_SQL;
 
 @Repository
 @RequiredArgsConstructor
