+++ src/main/java/gravit/code/userLeague/repository/custom/UserLeagueQueryRepository.java
@@ -1,0 +1,5 @@
+package gravit.code.userLeague.repository.custom;
+
+public interface UserLeagueQueryRepository extends LeagueRankingQueryRepository, MyLeagueProfileQueryRepository {
+
+}
