+++ src/main/java/gravit/code/userLeague/service/UserLeagueQueryService.java
@@ -3,20 +3,32 @@ package gravit.code.userLeague.service;
 import gravit.code.global.dto.response.SliceResponse;
 import gravit.code.global.exception.domain.CustomErrorCode;
 import gravit.code.global.exception.domain.RestApiException;
-import gravit.code.user.domain.UserRepository;
+import gravit.code.user.repository.UserRepository;
 import gravit.code.userLeague.dto.response.LeagueRankRow;
+import gravit.code.userLeague.dto.response.MyLeagueRankWithProfileResponse;
+import gravit.code.userLeague.repository.custom.UserLeagueQueryRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@RequiredArgsConstructor
 @Slf4j
-public class LeagueRankingQueryService {
+@RequiredArgsConstructor
+@Service
+public class UserLeagueQueryService {
 
-    private final LeagueRankingQueryRepository rankingQueryRepository;
     private final UserRepository userRepository;
+    private final UserLeagueQueryRepository userLeagueQueryRepository;
+
+    @Transactional(readOnly = true)
+    public MyLeagueRankWithProfileResponse getMyLeagueRankWithProfile(long userId) {
+
+        if(!userRepository.existsById(userId)){
+            throw new RestApiException(CustomErrorCode.USER_NOT_FOUND);
+        }
+
+        return userLeagueQueryRepository.findLeagueRankAndProfile(userId);
+    }
 
     @Transactional(readOnly = true)
     public SliceResponse<LeagueRankRow> findLeagueRanking(
@@ -24,7 +36,7 @@ public class LeagueRankingQueryService {
             int page
     ){
         int safePage = Math.max(0, page);
-        return rankingQueryRepository.findLeagueRanking(leagueId, safePage);
+        return userLeagueQueryRepository.findLeagueRanking(leagueId, safePage);
     }
 
     @Transactional(readOnly = true)
@@ -38,7 +50,6 @@ public class LeagueRankingQueryService {
             throw new RestApiException(CustomErrorCode.USER_NOT_FOUND);
         }
 
-        return rankingQueryRepository.findLeagueRankingByUser(userId, safePage);
+        return userLeagueQueryRepository.findLeagueRankingByUser(userId, safePage);
     }
-
 }
