+++ src/main/java/gravit/code/userLeague/service/UserLeagueService.java
@@ -7,9 +7,9 @@ import gravit.code.league.repository.LeagueRepository;
 import gravit.code.season.domain.Season;
 import gravit.code.season.service.SeasonService;
 import gravit.code.user.domain.User;
-import gravit.code.user.domain.UserRepository;
+import gravit.code.user.repository.UserRepository;
 import gravit.code.userLeague.domain.UserLeague;
-import gravit.code.userLeague.domain.UserLeagueRepository;
+import gravit.code.userLeague.repository.UserLeagueRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
